Recuperacao da 1 versao da classe produto java

+++ PastaFontesApp/Produto.java
@@ -1,21 +1,19 @@
-//Classe que guarda dados de produtos e serviços oferecidos no estacionamento
-//Exemplos de produtos e serviços: estacionar veiculo; lavar veiculo; trocar oleo
 
 public class Produto {
 
-	private int peso;
+	private double peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(int peso, double altura) {
+	public Produto(double peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public int getPeso() {
+	public double getPeso() {
 		return peso;
 	}
